refactor: structure of webhook sender folder
